Restructured response and input validation logic into their own dirs
