+++ CONTRIBUTORS.rst
@@ -1,0 +1,5 @@
+Project Maintainers
+-------------------
+
+* Sondre Lillebø Gundersen | `@sondrelg <https://github.com/sondrelg>`_
+* Jonas Krüger Svensson | `@JonasKs <https://github.com/JonasKs>`_
